+++ src/server.js
@@ -1,9 +1,9 @@
 import http from 'http';
+import util from 'util';
 import EventEmitter from 'events';
 import _ from 'underscore';
-import Promise from 'bluebird';
 import {
-    parseHostHeader, parseProxyAuthorizationHeader, parseUrl, redactParsedUrl,
+    parseHostHeader, parseProxyAuthorizationHeader, parseUrl, redactParsedUrl, nodeify,
 } from './tools';
 import HandlerForward from './handler_forward';
 import HandlerTunnelDirect from './handler_tunnel_direct';
@@ -427,10 +427,10 @@ export class Server extends EventEmitter {
     /**
      * Starts listening at a port specified in the constructor.
      * @param callback Optional callback
-     * @return {*}
+     * @return {(Promise|undefined)}
      */
     listen(callback) {
-        return new Promise((resolve, reject) => {
+        const promise = new Promise((resolve, reject) => {
             // Unfortunately server.listen() is not a normal function that fails on error,
             // so we need this trickery
             const onError = (err) => {
@@ -451,8 +451,9 @@ export class Server extends EventEmitter {
             this.server.on('error', onError);
             this.server.on('listening', onListening);
             this.server.listen(this.port);
-        })
-            .nodeify(callback);
+        });
+
+        return nodeify(promise, callback);
     }
 
     /**
@@ -503,7 +504,8 @@ export class Server extends EventEmitter {
         if (this.server) {
             const { server } = this;
             this.server = null;
-            return Promise.promisify(server.close).bind(server)().nodeify(callback);
+            const promise = util.promisify(server.close).bind(server)();
+            return nodeify(promise, callback);
         }
     }
 }