+++ test/anonymize_proxy.js
@@ -1,10 +1,10 @@
 const _ = require('underscore');
+const util = require('util');
 const { expect, assert } = require('chai');
 const proxy = require('proxy');
 const http = require('http');
 const portastic = require('portastic');
 const basicAuthParser = require('basic-auth-parser');
-const Promise = require('bluebird');
 const request = require('request');
 const express  = require('express');
 
@@ -81,7 +81,7 @@ before(() => {
 
 after(function () {
     this.timeout(5 * 1000);
-    if (proxyServer) return Promise.promisify(proxyServer.close).bind(proxyServer)();
+    if (proxyServer) return util.promisify(proxyServer.close).bind(proxyServer)();
 });